refactor(direct-control): extract unit operation postconditions
Move the synchronous unit-operation postcondition classifier group into packages/civ7-direct-control/src/play/operations/unit-postconditions.ts while keeping population and production postconditions plus wrapper-level request composition in the facade.

Record the new helper owner in the support modularization OpenSpec tasks and workstream records. This is local source-relocation proof only, not runtime/live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
